OM-236 | Add configuration for CI pipelines
Changed the Docker configs based on what was configured in the
kukkuu-ui application. These should now include a nginx config
for the health check endpoints. Added a yaml file for Gitlab.

The values in the configs are mostly guesses. We'll see how this
plays out in the actual CI build.