Centralize Apollo and Sentry integration
Previously the Apollo and Sentry integration was done multiple times at
multiple different places in the code. This commit centralizes this
logic by logging errors into Sentry by using the 'onError' field in
Apollo boost's client.

This centralized error handling adds more context to the data we pass
on to Sentry. This makes debugging easier by allowing us to see the
name of the query that caused an error, for instance. This kind of
context couldn't have been added without centralization.

This change also implements a centralized error ignoring mechanism.
We know that we do not want to pass some errors to Sentry, because the
errors we receive are expected. Such error often include for instance
permission related errors. Now such errors can be ignored globally,
making it less likely that we forget to ignore them in some view.

Errors can be ignore or consumed. Ignored errors won't be sent to
Sentry, but they will be sent to consumers like 'useQuery' hooks. When
an error is consumed, it won't be passed on to consumers, in which case
'useQuery' hook's 'onError' callback would not be called.

Note that when consuming en error, the entire error object is consumed.
This means that if one error in an error object containing two errors
is consumed, both errors are consumed.

Note that global error consuming should be used scarcely. It should be
reserved for those errors of whose exclusion the view layer clearly
benefits from.