Reorder GraphQL calls for creating youth membership
Due to open-city-profile deprecating some old behavior, the order
of the calls is changed when registering a new member to:

1. Create Helsinki profile
2. Add connection to Youth membership service
3. Create Youth profile

The last call eventually also calls the Helsinki profile in the youth membership
backend. The service connection needs to be created into Helsinki profile
before the youth membership profile creation call is made.

Refs YM-512